Moved urlgraph to new file.
